resolve conflic
