+++ src/pages/Home/components/News/index.js
@@ -1,38 +1,42 @@
 import React from 'react';
 import Typography from 'components/Typography';
+import Card, { CardsLoading } from 'components/Card';
 import SubmitReportButton from 'components/SubmitReportButton';
 import Button from 'components/Button';
-import Card from 'components/Card';
 import { Wrapper, ButtonsWrapper, WrapperCards } from './index.view';
 import { useReport } from 'contexts/report.context';
 
 const News = () => {
-  const { reliableReports, viewReportDetail } = useReport();
+  const { fetching, reliableReports, viewReportDetail } = useReport();
   return (
     <Wrapper>
       <Typography variant="body" weight="normal">
         แหล่งข่าวน่าเชื่อถือ
       </Typography>
       <WrapperCards>
-        {reliableReports.map(report => {
-          return (
-            <Card
-              key={report.id}
-              report={report}
-              onClick={viewReportDetail}
-              image={report.image}
-              header={report.header}
-              content={report.content}
-              avatar={report.avatar}
-              reference={report.link}
-              location={report.location}
-              another={report.header.another}
-              id={report.id}
-              dislikes={report.dislikes}
-              likes={report.likes}
-            />
-          );
-        })}
+        {fetching ? (
+          <CardsLoading />
+        ) : (
+          reliableReports.map(report => {
+            return (
+              <Card
+                key={report.id}
+                report={report}
+                onClick={viewReportDetail}
+                image={report.image}
+                header={report.header}
+                content={report.content}
+                avatar={report.avatar}
+                reference={report.link}
+                location={report.location}
+                another={report.header.another}
+                id={report.id}
+                dislikes={report.dislikes}
+                likes={report.likes}
+              />
+            );
+          })
+        )}
       </WrapperCards>
 
       <ButtonsWrapper>
