+++ src/pages/Report/Full/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
 import { List } from 'antd';
-import CardCustom from 'components/Card';
+import CardCustom, { CardsLoading } from 'components/Card';
 import { GoBackHeader } from 'components/BarNavigation/navigation';
+
 import { useReport } from 'contexts/report.context';
 
 const Container = styled.div`
@@ -12,32 +13,36 @@ const Container = styled.div`
 `;
 
 const FullReport = () => {
-  const { reports, viewReportDetail } = useReport();
+  const { fetching, reports, viewReportDetail } = useReport();
 
   return (
     <>
       <GoBackHeader mxwidth="680px" />
       <Container>
         <List>
-          {reports.reverse().map(report => {
-            return (
-              <CardCustom
-                key={report.id}
-                report={report}
-                onClick={viewReportDetail}
-                image={report.image}
-                header={report.header}
-                content={report.content}
-                avatar={report.avatar}
-                reference={report.link}
-                location={report.location}
-                another={report.header.another}
-                id={report.id}
-                dislikes={report.dislikes}
-                likes={report.likes}
-              />
-            );
-          })}
+          {fetching ? (
+            <CardsLoading />
+          ) : (
+            reports.reverse().map(report => {
+              return (
+                <CardCustom
+                  key={report.id}
+                  report={report}
+                  onClick={viewReportDetail}
+                  image={report.image}
+                  header={report.header}
+                  content={report.content}
+                  avatar={report.avatar}
+                  reference={report.link}
+                  location={report.location}
+                  another={report.header.another}
+                  id={report.id}
+                  dislikes={report.dislikes}
+                  likes={report.likes}
+                />
+              );
+            })
+          )}
         </List>
       </Container>
     </>
